+++ tests/conftest.py
@@ -29,7 +29,7 @@ def uncorrelated_model():
 
 @pytest.fixture
 def simple_limit_state():
-    """Basic limit state function: g = X1 - X2."""
+    """Basic limit-state function: g = X1 - X2."""
 
     def lsf(R, S):
         return R - S